refactor: update import paths to use constants from the new location
