+++ src/components/theme/Views/DatabaseItemView.jsx
@@ -12,10 +12,7 @@ import {
   flourishDataprotection,
   getDataSrcFromEmbedCode,
 } from '@eeacms/volto-cca-policy/helpers/flourishUtils';
-import {
-  VIDEO,
-  CONTENT_TYPE_LABELS,
-} from '@eeacms/volto-cca-policy/helpers/Constants';
+import { VIDEO, CONTENT_TYPE_LABELS } from '@eeacms/volto-cca-policy/constants';
 import {
   HTMLField,
   ReferenceInfo,
